Moving the onEnd logic in the common base class.
