+++ hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/migration/TimeSensitiveBalanceMigration.java
@@ -18,6 +18,8 @@ package com.hedera.mirror.importer.migration;
 
 import com.hedera.mirror.common.domain.balance.AccountBalanceFile;
 import com.hedera.mirror.common.domain.transaction.RecordFile;
+import com.hedera.mirror.importer.exception.ImporterException;
+import com.hedera.mirror.importer.parser.balance.BalanceStreamFileListener;
 import com.hedera.mirror.importer.repository.AccountBalanceFileRepository;
 import com.hedera.mirror.importer.repository.RecordFileRepository;
 import java.io.IOException;
@@ -27,7 +29,8 @@ import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.transaction.support.TransactionSynchronization;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
 
-abstract class TimeSensitiveBalanceMigration extends RepeatableMigration implements TransactionSynchronization {
+abstract class TimeSensitiveBalanceMigration extends RepeatableMigration
+        implements BalanceStreamFileListener, TransactionSynchronization {
 
     private final AccountBalanceFileRepository accountBalanceFileRepository;
     private final RecordFileRepository recordFileRepository;
@@ -71,6 +74,17 @@ abstract class TimeSensitiveBalanceMigration extends RepeatableMigration impleme
                 .isPresent()) {
             TransactionSynchronizationManager.registerSynchronization(this);
             doMigrate();
+        }
+    }
+
+    @Override
+    public void onEnd(AccountBalanceFile accountBalanceFile) throws ImporterException {
+        try {
+            initializeBalanceTimeSensitively(accountBalanceFile);
+        } catch (IOException e) {
+            log.error(
+                    "Error executing the migration again after consensus_timestamp {}",
+                    accountBalanceFile.getConsensusTimestamp());
         }
     }
